move all processing function into processing.py
