--- data_processing/preprocessing.py
@@ -1,40 +1,0 @@
-# Copyright 2016-2019 The Van Valen Lab at the California Institute of
-# Technology (Caltech), with support from the Paul Allen Family Foundation,
-# Google, & National Institutes of Health (NIH) under Grant U24CA224309-01.
-# All rights reserved.
-#
-# Licensed under a modified Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.github.com/vanvalenlab/kiosk-data-processing/LICENSE
-#
-# The Work provided may be used for non-commercial academic purposes only.
-# For any other use of the Work, including commercial use, please contact:
-# vanvalenlab@gmail.com
-#
-# Neither the name of Caltech nor the names of its contributors may be used
-# to endorse or promote products derived from this software without specific
-# prior written permission.
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ============================================================================
-"""Functions for pre-processing predictions from tf-serving"""
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-
-def noramlize(image):
-    """Normalize image data by dividing by the maximum pixel value
-    # Arguments:
-        image: numpy array of image data
-    # Returns:
-        normal_image: normalized image data
-    """
-    normal_image = (image - image.mean()) / image.std()
-    return normal_image
